use transaction

+++ Asp In Action/Services/CostControl/Entity/Transaction.cs
@@ -4,7 +4,7 @@ namespace Asp_In_Action.Services.CostControl.Entity
 {
     public class Transaction
     {
-        public int Id { get; set; }
+        public ulong Id { get; set; }
         public TransactionType Type { get; set; }
         public DateTime Date { get; set; }
         public string Description { get; set; }
